interaction backend

+++ components/global/Interaction.tsx
@@ -1,4 +1,4 @@
-import { motion } from "framer-motion";
+import { AnimatePresence, motion } from "framer-motion";
 import {
   ExpandCurveIcon,
   TinyCurveIcon,
@@ -6,6 +6,7 @@ import {
   AddIcon,
   DeleteIcon,
   ReportIcon,
+  ExpandIcon,
 } from "@/components/icons";
 import React, { useCallback } from "react";
 import { ArtifactExtended } from "@/types/globalTypes";
@@ -16,13 +17,21 @@ import { toast } from "sonner";
 import { createFlag, deleteEntry } from "@/lib/helper/artifact";
 
 import { FlagType } from "@prisma/client";
+import { useArtifact } from "@/hooks/usePage";
 
-// Define the animation variants
 const dotVariants = {
   hidden: { opacity: 0, scale: 0.75 },
   visible: {
     opacity: 1,
     scale: 1,
+    transition: {
+      type: "spring",
+      damping: 49,
+      stiffness: 400,
+    },
+  },
+  hovered: {
+    scale: 1.25,
     transition: {
       type: "spring",
       damping: 49,
@@ -41,7 +50,6 @@ const containerVariants = {
   },
 };
 
-// Create a styled dot component
 const Dot = () => {
   return (
     <motion.div
@@ -59,9 +67,17 @@ export const Interaction = ({ artifact }: InteractionProps) => {
   const { playContent, setSelectedFormSound } = useSoundContext();
   const { setIsVisible, user } = useInterfaceContext();
   const { setExpandInput } = useNavContext();
+  const { handleSelectArtifact } = useArtifact();
+
+  const [isDeleting, setIsDeleting] = React.useState(false);
+  const [isFlagging, setIsFlagging] = React.useState(false);
 
   const [isHovered, setIsHovered] = React.useState(false);
   const sound = artifact.appleData;
+
+  const handleOpen = () => {
+    handleSelectArtifact(artifact);
+  };
 
   const handlePlayContent = async () => {
     playContent(sound.id, sound.type);
@@ -73,26 +89,48 @@ export const Interaction = ({ artifact }: InteractionProps) => {
     setExpandInput(true);
   };
 
-  const handleDelete = useCallback(async (artifactId: string) => {
-    // Wrap the deleteEntry call inside toast.promise to manage toast notifications
-    toast.promise(
-      deleteEntry(artifactId).then(() => {
-        // For example, updating the UI or state to reflect the deletion
-      }),
-      {
-        loading: "Deleting...",
-        success: "Deletion successful!",
-        error: "Error deleting artifact",
-      },
-    );
-  }, []);
+  const handleDelete = useCallback(
+    async (artifactId: string) => {
+      if (!isDeleting) {
+        setIsDeleting(true);
+        setTimeout(() => setIsDeleting(false), 3000); // 3 seconds to confirm
+        return;
+      }
 
-  const handleFlag = () => {
-    toast.promise(createFlag(artifact.id, FlagType.artifact, user!.id), {
-      loading: "Flagging...",
-      success: "Content flagged successfully!",
-      error: "Error flagging content",
-    });
+      toast.promise(
+        deleteEntry(artifactId).then(() => {
+          setIsDeleting(false);
+          // For example, updating the UI or state to reflect the deletion
+        }),
+        {
+          loading: "Deleting...",
+          success: "Deletion successful!",
+          error: "Error deleting artifact",
+        },
+      );
+    },
+    [isDeleting],
+  );
+
+  const handleFlag = async () => {
+    if (!isFlagging) {
+      setIsFlagging(true);
+      setTimeout(() => setIsFlagging(false), 3000);
+      return;
+    }
+
+    try {
+      const response = await createFlag(
+        artifact.id,
+        FlagType.artifact,
+        user!.id,
+      );
+      console.log("Content flagged successfully!", response);
+      setIsFlagging(false);
+    } catch (error) {
+      console.error("Error flagging content", error);
+      setIsFlagging(false);
+    }
   };
 
   return (
@@ -115,6 +153,7 @@ export const Interaction = ({ artifact }: InteractionProps) => {
 
       {/* System */}
       <motion.div
+        layout
         className="flex origin-top-right flex-row-reverse items-center justify-center gap-2 pr-2"
         variants={containerVariants}
         animate={isHovered ? "visible" : "hidden"}
@@ -122,23 +161,52 @@ export const Interaction = ({ artifact }: InteractionProps) => {
         {Array.from({ length: 2 }).map((_, index) => (
           <Dot key={index} />
         ))}
-        <motion.div
-          onClick={() => handleDelete(artifact.id)}
-          variants={dotVariants}
-          className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
-        >
-          <DeleteIcon />
-        </motion.div>
-        <motion.div
-          className="bg-gray4 h-1 w-1 rounded-full"
-          variants={dotVariants}
-        />
-        <motion.div
-          onClick={handleFlag}
-          variants={dotVariants}
-          className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
-        >
-          <ReportIcon />
+
+        {/* Delete button */}
+        {user?.id === artifact.author.id && (
+          <>
+            <motion.div
+              whileHover={{
+                scale: 1.25,
+              }}
+              whileTap={{
+                scale: 0.75,
+              }}
+            >
+              <motion.div
+                layout
+                onClick={() => handleDelete(artifact.id)}
+                variants={dotVariants}
+                className={`flex items-center justify-center rounded-full bg-[#F20000] bg-opacity-25 p-2 text-xs font-medium uppercase leading-[8px] text-[#F20000]`}
+              >
+                <AnimatePresence>
+                  {isDeleting ? (
+                    <div>archive</div>
+                  ) : (
+                    <DeleteIcon color={"#F20000"} />
+                  )}
+                </AnimatePresence>
+              </motion.div>
+            </motion.div>
+
+            <motion.div
+              className="bg-gray4 h-1 w-1 rounded-full"
+              variants={dotVariants}
+            />
+          </>
+        )}
+
+        {/* Flag button */}
+        <motion.div whileHover={{ scale: 1.25 }} whileTap={{ scale: 0.75 }}>
+          <motion.div
+            onClick={handleFlag}
+            variants={dotVariants}
+            className={`flex items-center justify-center rounded-full bg-[#FF5200] bg-opacity-25 p-2 text-xs font-medium uppercase leading-[8px] text-[#F20000]`}
+          >
+            <AnimatePresence>
+              {isFlagging ? <div>flag</div> : <ReportIcon color={"#F20000"} />}
+            </AnimatePresence>
+          </motion.div>
         </motion.div>
       </motion.div>
 
@@ -152,25 +220,42 @@ export const Interaction = ({ artifact }: InteractionProps) => {
           {Array.from({ length: 9 }).map((_, index) => (
             <Dot key={index} />
           ))}
-          <motion.div
-            onClick={handleCreate}
-            variants={dotVariants}
-            className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
-          >
-            <AddIcon />
+
+          {/* Create button */}
+          <motion.div whileHover={{ scale: 1.25 }} whileTap={{ scale: 0.75 }}>
+            <motion.div
+              onClick={handleCreate}
+              variants={dotVariants}
+              className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
+            >
+              <AddIcon />
+            </motion.div>
           </motion.div>
+
           <motion.div
             className="bg-gray4 h-1 w-1 rounded-full"
             variants={dotVariants}
           />
-          <motion.button
-            onClick={handlePlayContent}
-            variants={dotVariants}
-            className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
+
+          {/* Play button */}
+          <motion.div
+            whileHover={{
+              scale: 1.25,
+            }}
+            whileTap={{
+              scale: 0.75,
+            }}
           >
-            <PlayIcon />
-          </motion.button>
+            <motion.button
+              onClick={handlePlayContent}
+              variants={dotVariants}
+              className={`bg-gray4 flex h-8 w-8 items-center justify-center rounded-full`}
+            >
+              <PlayIcon />
+            </motion.button>
+          </motion.div>
         </motion.div>
+
         <motion.div
           animate={{
             x: !isHovered ? -12 : 0,
@@ -178,8 +263,20 @@ export const Interaction = ({ artifact }: InteractionProps) => {
             scale: !isHovered ? 0 : 1,
             opacity: !isHovered ? 0 : 1,
           }}
-          className={`h-8 w-8 origin-top-left`}
+          whileTap={{
+            scale: 0.75,
+          }}
+          onClick={handleOpen}
+          className={`relative h-8 w-8 origin-top-left`}
         >
+          <motion.div
+            whileHover={{
+              scale: 1.25,
+            }}
+            className={`absolute -left-[11px] -top-[11px]`}
+          >
+            <ExpandIcon />
+          </motion.div>
           <TinyCurveIcon color={"#F4F4F4"} />
         </motion.div>
       </div>
